+++ DatadogRUM/Sources/Instrumentation/Actions/RUMActionsHandler.swift
@@ -26,6 +26,7 @@ internal final class RUMActionsHandler: RUMActionsHandling {
     /// Convenience initializer for iOS
     convenience init(
         dateProvider: DateProvider,
+        heatmapIdentifierRegistry: any HeatmapIdentifierRegistry,
         uiKitPredicate: UITouchRUMActionsPredicate?,
         swiftUIPredicate: SwiftUIRUMActionsPredicate?,
         swiftUIDetector: SwiftUIComponentDetector?
@@ -39,6 +40,7 @@ internal final class RUMActionsHandler: RUMActionsHandling {
             dateProvider: dateProvider,
             eventCommandsFactory: UITouchCommandFactory(
                 dateProvider: dateProvider,
+                heatmapIdentifierRegistry: heatmapIdentifierRegistry,
                 uiKitPredicate: uiKitPredicate,
                 swiftUIPredicate: swiftUIPredicate,
                 swiftUIDetector: swiftUIDetector