+++ DatadogRUM/Tests/Instrumentation/Actions/RUMActionsHandlerTests.swift
@@ -23,6 +23,7 @@ class RUMActionsHandlerTests: XCTestCase {
         let handler =
         RUMActionsHandler(
             dateProvider: dateProvider,
+            heatmapIdentifierRegistry: HeatmapIdentifierRegistryMock(),
             uiKitPredicate: uiKitPredicate,
             swiftUIPredicate: swiftUIPredicate,
             swiftUIDetector: SwiftUIComponentFactory.createDetector()